requirements.txt
asttokens==3.0.1
branca==0.8.2
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
comm==0.2.3
contourpy==1.3.2
cryptography==46.0.7
cycler==0.12.1
db-dtypes==1.5.1
debugpy==1.8.20
decorator==5.2.1
dotenv==0.9.9
exceptiongroup==1.3.1
executing==2.2.1
flatbuffers==25.9.23
folium==0.20.0
fonttools==4.62.1
GeoAlchemy2==0.19.0
geographiclib==2.1
geopandas==1.1.2
geopy==2.4.1
google-api-core==2.30.3
google-auth==2.49.2
google-cloud-bigquery==3.41.0
google-cloud-bigquery-storage==2.37.0
google-cloud-core==2.5.1
google-crc32c==1.8.0
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
greenlet==3.4.0
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
idna==3.13
ipykernel==7.2.0
ipython==8.39.0
jedi==0.19.2
jh2==5.0.11
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.5.0
mapclassify==2.8.1
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
networkx==3.4.2
niquests==3.18.6
numpy==2.2.6
openmeteo_requests==1.7.5
openmeteo_sdk==1.26.0
packaging==26.1
pandas==2.3.3
parso==0.8.6
pexpect==4.9.0
pillow==12.2.0
platformdirs==4.9.6
prompt_toolkit==3.0.52
proto-plus==1.27.2
protobuf==6.33.6
psutil==7.2.2
psycopg==3.3.3
psycopg-binary==3.3.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==24.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.20.0
pyogrio==0.12.1
pyparsing==3.3.2
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
pyzmq==27.1.0
qh3==1.7.3
requests==2.33.1
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
shapely==2.1.2
six==1.17.0
SQLAlchemy==2.0.49
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.5.5
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
urllib3-future==2.19.909
wassima==2.0.6
wcwidth==0.6.0
xyzservices==2026.3.0
